Install `uv` to temp dir during bootstrap script

+++ bootstrap.sh
@@ -4,11 +4,11 @@ set -eu
 # Bootstrap: curl -LsSf https://raw.githubusercontent.com/mohdfareed/machine/main/bootstrap.sh | sh
 ROOT="${MC_ROOT:-$HOME/.machine}"
 UV="uv"
+TMP_DIR=""
 
-# ensure git is available
-if ! command -v git >/dev/null 2>&1; then
-    echo "Error: git is not installed. Please install git and try again."
-    exit 1
+# clone repo if needed
+if ! [ -d "$ROOT/.git" ]; then
+    git clone https://github.com/mohdfareed/machine.git "$ROOT"
 fi
 
 # ensure uv is available
@@ -18,16 +18,14 @@ if ! command -v uv >/dev/null 2>&1; then
     export UV_NO_MODIFY_PATH=1
     export UV_DISABLE_UPDATE=1
 
+    # install uv to temp dir
     curl -LsSf https://astral.sh/uv/install.sh | sh
     UV="$TMP_DIR/uv"
-fi
-
-# clone repo if needed
-if ! [ -d "$ROOT/.git" ]; then
-    git clone https://github.com/mohdfareed/machine.git "$ROOT"
+    bin="$("$UV" tool dir --bin)"
+    export PATH="$bin:$PATH"
 fi
 
 # install the cli tool
 "$UV" tool install "$ROOT" --force
-rm -rf "$TMP_DIR"
+[ -n "$TMP_DIR" ] && rm -rf "$TMP_DIR" # clean up
 echo "Done. Run 'mc --help' to get started."
